agents/pyproject.toml
[tool.poetry]
name = "agent-workflow-backend"
version = "0.1.0"
description = "Agent Workflow Backend API"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
motor = "^3.7.0"
python-dotenv = "^1.1.0"
jwt = "^1.3.1"
passlib = "1.7.4"
python-multipart = "^0.0.20"
bcrypt = "4.0.1"
python-jose = "^3.4.0"
openai = "^1.12.0"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 